app/requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt
gunicorn==20.1.0
psycopg2-binary==2.9.3
flake8==3.9.2
